backend/pyproject.toml
[tool.poetry]
name = "test-unwind"
version = "0.1.0"
description = "Test task for Unwind"
authors = ["Andrey Chernikov"]
license = "MIT"
readme = "README.md"
packages = [{include = "test_unwind"}]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^3.2"
djangorestframework = "^3.13.1"
drf-spectacular = "^0.23.1"
python-dotenv = "^0.21.0"
django-debug-toolbar = "^3.6.0"
django-cors-headers = "^3.13.0"
django-filter = "^22.1"
google-api-core = "^2.10.0"
google-api-python-client = "^2.60.0"
google-auth-oauthlib = "^0.5.2"
oauth2client = "^4.1.3"
requests = "^2.28.1"
psycopg2-binary = "^2.9.3"
celery = "^5.2.7"
redis = "^4.3.4"
xmltodict = "^0.13.0"
gunicorn = "^20.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
src_paths = ["project", "orders"]
